feat:新增限流,流量兜底,以及获取,创建商品接口

+++ pay/internal/repository/pay.go
@@ -3,14 +3,14 @@ package repository
 import "gorm.io/gorm"
 
 type Pay struct {
-	PayID  int `gorm:"primaryKey" json:"id"`
-	UserID int `json:"userId"`
-	ItemID int `json:"itemId"`
-	Amount int `json:"amount"`
+	PayID  int     `gorm:"primaryKey" json:"id"`
+	UserID int     `json:"userId"`
+	ItemID int     `json:"itemId"`
+	Amount float32 `json:"amount"`
 }
 
 type PayRepository interface {
-	CreatePay(userID int, itemID int, amount float64) error
+	CreatePay(userID int, itemID int, amount float32) error
 }
 
 type PayRepositoryImpl struct {
@@ -21,7 +21,7 @@ func NewPayRepositoryImpl(db *gorm.DB) *PayRepositoryImpl {
 	return &PayRepositoryImpl{db: db}
 }
 
-func (p *PayRepositoryImpl) CreatePay(userID int, itemID int, amount int) error {
+func (p *PayRepositoryImpl) CreatePay(userID int, itemID int, amount float32) error {
 	pay := &Pay{
 		UserID: userID,
 		ItemID: itemID,
